7.10.1, belated 7.8.4, move to new upstream PPA
The deb.haskell.org repository has been abandoned in this version update
since it was not up to date and has been taken offline after a recent
compromise.  We have settled on sticking with a Debian base and
utilizing a PPA maintained by an active GHC committer.  The new PPA
packages are based on the original source, but have been actively
maintained.  The 7.8.4 release, which is the last in the 7.8 series is
being included even though it should have gone out months ago when it
was initially released and is now basically dead-end.  There are still a
number of libraries that fail on 7.10 at this point.  With the next
version bump we plan to properly drop the 7.8 tag.