Update registry:latest to point to registry 2.x
The registry:latest tag currently points to registry 0.9.1. It has been
pointing to this old version for awhile to avoid breaking users
following outdated legacy registry documentation. Registry 2.x is
well-established now, and the time has come to change the latest tag.
Change it so that registry:latest is synonymous with registry:2.

Remove old 0.8.1 and 0.9.1 tags from the file.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>