bashbrew: Implement parallel push support
All docker push commands will be run in background to speed up the whole
process. There is a maximum number of running background jobs (Default
to 20), if it reaches the limit then put the main thread to sleep
(Default to 2 seconds) and check the number of running bg jobs again. We
can tune these numbers later.

Change-type:patch
Signed-off-by: Trong Nghia Nguyen <nghiant2710@gmail.com>