Remove Sentry image
Sentry now requires multiple services to be set up and having a single official image, especially with the lack of cooperation from the Docker team for updates (see getsentry/docker-sentry#211) does not make sense. This page and image is now dragging us back as we want people to use the newer versions of Sentry.

Ideally, the old page would redirect to https://develop.sentry.dev/self-hosted/ or at least https://hub.docker.com/r/getsentry/sentry but it is okay if this is not possible.

/cc @mattrobenolt @benvinegar @dcramer