Add `redhat/ubi9` and `redhat/ubi8{,-minimal}` base image exceptions
These complement our existing/initial `redhat/ubi9-minimal` exception such that official images can now be based on UBI8 or UBI9 (in either their standard or "minimal" variants).

In evaluating which variants to add to our list of exceptions from https://hub.docker.com/u/redhat, we've chosen to exclude `init` ("one concern per container") and `micro` ("uses the package manager on the underlying host").

Additionally, these exceptions are all explicitly/intentionally on `:latest` because of the unique repository scheme employed upstream (one repository per major version and variant) and because one of the main benefits to the official images program is that image rebuilds happen automatically based on base image updates, so an image pinning to a minor release of UBI8 or UBI9 would not be acceptable (as it would not be something we could automatically rebuild when the base image has updates for security fixes, etc).

The intended `FROM` value for interested image maintainers is something like `FROM redhat/ubi9-minimal` (and our build infrastructure will apply our pinning automatically so that we can appropriately trigger rebuilds).