Update Debian (esp. for 12.1)
Unfortunately, `riscv64` is in a bit of a weird place right at the moment (https://lists.debian.org/debian-riscv/2023/07/msg00053.html) as it has been officially accepted into the archive (🎉), but is still in the process of being bootstrapped (😅), so the archive doesn't have much `riscv64` in it yet (just the in-progress bootstrapping) but ports also isn't likely to be updated further.  For now, this still points to ports, and we'll reevaluate at the next update when the bootstrapping has had a bit longer to progress.