Test that setuptools and wheel are installed (or not)
In Python 3.12+, setuptools and wheel should be removed (but should be present in older versions).

Technically this should be tested on Python 2 as well, but I opted not to because the interfaces for doing so are different and that only exists in a single variant of PyPy (that frankly I don't care too much about verifying the "correctness" of as much).