Improve `toc.sh` to support more of GitHub's edge cases
Most notably, this allows emoji in headers to generate the correct anchors.

See https://gist.github.com/tianon/75e267d9137b1c2978031b66b3a98987 for a bunch of test cases I worked through this with (and found bugs in GitHub's implementation while reverse engineering it, which is fun, and quirks of the library I was copying that aren't correct for the GitHub implementation).

I've also verified that this causes no changes to the TOC here or over in the docs repo (as expected).