maridadb: 2025 Q3 release
Notabely: MariaDB-10.6 is now based on jammy (24.04)
due to Focal 20.04 EOL

12.0 is now latest.

With UBI10 being stable, the 12.0+ series is based on
UBI10 for the ubi series.

The ubi images have now a FIP enable by default
profile for connections. SQL functions like MD5/SHA1
are still available as they are explictly requested
of the OpenSSL library